+++ tests/testthat/test-estimate_contrasts_methods.R
@@ -173,11 +173,10 @@ test_that("estimate_contrasts - random effects, single by/contrast", {
     estimate_contrasts(model, contrast = "gear"),
     regex = "Could not calculate"
   )
-  expect_message(
+  expect_silent(
     {
       estim <- estimate_relation(model, by = "gear")
-    },
-    regex = "Standard errors are probably not reliable"
+    }
   )
   out <- estimate_contrasts(estim, contrast = "gear")
   expect_identical(dim(out), c(3L, 8L))